In this modified code:
I've added logging statements to the run_swarms method to provide feedback to the user about the progress of the tasks.

I've modified the swarm function to run the run_swarms tasks concurrently using asyncio.wait. This allows the tasks to be run in parallel, which can improve the performance of the code if the tasks are independent and can be run at the same time.


Former-commit-id: 24e4dc021f8ef43f6dd77a5cac2cc7b9125369f4